+++ packages/client/src/errors/index.ts
@@ -1,1 +1,2 @@
 export * from './SfuJoinError';
+export * from './SfuTimeoutError';