+++ packages/react-native-callingx/ios/Callingx.mm
@@ -169,6 +169,9 @@ RCT_EXPORT_MODULE(Callingx)
   WebRTCModule *webrtcModule = [self.moduleRegistry moduleForName:"WebRTCModule"];
   _moduleImpl.webRTCModule = webrtcModule;
 
+  // Must run after webRTCModule injection: getAudioDeviceModule() depends on it.
+  [_moduleImpl wireEngineSubscription];
+
   self.callKeepCallController = _moduleImpl.callKeepCallController;
   self.callKeepProvider = _moduleImpl.callKeepProvider;
 }
@@ -559,11 +562,11 @@ RCT_EXPORT_METHOD(fulfillEndCallAction:(NSString *)callId didFail:(BOOL)didFail)
 
 #ifdef RCT_NEW_ARCH_ENABLED
 - (void)log:(NSString *)message level:(NSString *)level {
-  NSLog(@"[Callingx][log] %@, %@", message, level);
+  [CallingxLogBridge js:message level:level];
 }
 #else
 RCT_EXPORT_METHOD(log:(NSString *)message level:(NSString *)level) {
-  NSLog(@"[Callingx][log] %@, %@", message, level);
+  [CallingxLogBridge js:message level:level];
 }
 #endif